removing tons of compiler warnings! please try not to check in code that gives a warning when compiled. this is bad practice. i am guilty of committing this crime as well, of course.
git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1737 e4a34bab-3947-0410-866c-ab59ef520b34

+++ kamping/KChamonixKAmpSite.cxx
@@ -610,7 +610,7 @@ set<int>& KChamonixKAmpSite::GetHeatPulseStampWidths(KRawBoloPulseRecord *pRaw)
   for(int i = 0; i < bolo->GetNumPulseRecords(); i++){
     KRawBoloPulseRecord *p = bolo->GetPulseRecord(i);
     if(p->GetIsHeatPulse())
-      stampwidths.insert( (p->GetHeatPulseStampWidth()) );  
+      stampwidths.insert( (int)(p->GetHeatPulseStampWidth()) );  
   }
 
   fHeatPulseStampWidths[pRaw->GetChannelName()] = stampwidths;
