+++ kpta/KOrderFilter.h
@@ -23,8 +23,8 @@ public:
   
   virtual void SetOrder(unsigned int a){fOrder = a;}  
   virtual unsigned int GetOrder(void) const {return fOrder;}
-  virtual void SetInitOutputValue(double a){fInitOutputValue = a; fSetInitOutputValue = true;}
-  virtual double GetInitOutputValue(void){return fInitOutputValue;}
+  virtual void SetInitOutputValue(int a){fInitOutputValue = a; fSetInitOutputValue = true;}
+  virtual int GetInitOutputValue(void){return fInitOutputValue;}
   virtual bool IsSetInitOutputValue(void){return fSetInitOutputValue;}
   virtual void SetUseInitOutputValue(bool a){fSetInitOutputValue = a;}
   
@@ -32,7 +32,7 @@ public:
   
 private:
   unsigned int fOrder;
-  double fInitOutputValue;
+  int fInitOutputValue;
   bool fSetInitOutputValue;
   //private methods
   void InitializeMembers(void);
